backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
urllib3==2.2.3
